Sync final newlines in source code
This patch syncs final newlines in source code files:
- trims multiple final newlines into one
- adds one final newline where it was missing

According to POSIX, a line is a sequence of zero or more non-' <newline>'
characters plus a terminating '<newline>' character. [1] Files should
normally have at least one final newline character.

C89 standard [2] and above [3] mention a final newline:
"A source file that is not empty shall end in a new-line character,
which shall not be immediately preceded by a backslash character."

Newline characters:
- LF (\n) (*nix and Mac, default)
- CRLF (\r\n) (Windows)
- CR (\r) (old Mac, obsolete)

Although it is not mandatory for all files to have this issue fixed, a
more consistent and homogeneous approach brings less of commit
differences issues and a better development experience in certain text
editors and IDEs.

[1] http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_206
[2] https://port70.net/~nsz/c/c89/c89-draft.html#2.1.1.2
[3] https://port70.net/~nsz/c/c99/n1256.html#5.1.1.2