Added method getState()
When a MySql connection object is pulled from a connection pool, there is currently no "cheap" way to check if the connection is really still there before it is injected to the application side. Being able to check SW_MYSQL_STATE_CLOSED first is very efficient and the infrastructure for this is already in place in current code.