Fix truncated cookie
Previously, Max-Age was added to the cookie, but cookie_size wasn't
updated to account for the additional string size. This lead to the data
being truncated in cases where everything was set.

This change updates the cookie_size based on the length of max_age and
then increases the size of cookie to the increased cookie_size.