Fixing errno after wait
There is a bug in Swoole's event loop that causes warning messages for things that shouldn't be warnings.
The line "if (errno > 0 && errno != EINTR)" is expecting the global variable "errno" to be set by wait system call in the inline function "wait_process", but instead errno is getting overwritten by the "spawn_task_worker" or "spawn_event_worker methods" of the server.