AC_PROG_CC_C99 is obsolete with autoconf >= 2.70 (#5254)
To make sure that compiler supports C99 before Autoconf 2.69, this was
needed. But with Autoconf 2.70 and later the macro is obsolete because
the checks are done in AC_PROG_CC and warnings are emitted when building
configure script on newer Autoconf versions.

Once PHP will require Autoconf 2.70 or newer, this can be removed. The
AC_PROG_CC is added by phpize.m4 from the php-src repository.